PUBDEV-7037: added Activatable with ability to decide itself
