+++ h2o-core/src/main/java/hex/grid/filter/Activatable.java
@@ -7,6 +7,7 @@ public interface Activatable<MP extends Model.Parameters> {
   /**
    * Provides a way to set implementation into a state when it is considered to be already used. Useful for stateful implementations 
    * as stateless ones do not have any state.
+   * @param globalActivate provides global decision taking into account all {@link FilterFunction}
    */
-  void activate(MP permutation);
+  void activate(boolean globalActivate, MP permutation);
 }
